Fix Next route export by moving message policy helper
