+++ apps/web/app/api/messages/route.ts
@@ -17,37 +17,7 @@ import { validateAttachments, validateAttachmentContent } from "@/lib/attachment
 import { MESSAGE_PROJECTION, withReplyTo, type ServerSupabaseClient } from "@/lib/messages/hydration"
 import { parsePostMessageRequestBody, type MessageAttachment, type PostMessageRequestBody } from "@/lib/messages/validators"
 import { enqueueAttachmentScans } from "@/lib/attachment-malware"
-
-
-
-type SupportedMessageChannelType = "text" | "announcement" | "forum" | "media"
-
-export function validateChannelTypeMessagePolicy({
-  channelType,
-  hasSendPermission,
-  content,
-  attachments,
-}: {
-  channelType: SupportedMessageChannelType
-  hasSendPermission: boolean
-  content?: string
-  attachments: MessageAttachment[]
-}) {
-  if (!hasSendPermission) {
-    return { error: "Missing SEND_MESSAGES permission", status: 403 }
-  }
-
-
-  if (channelType === "media" && attachments.length === 0) {
-    return { error: "Media channels require at least one attachment.", status: 400 }
-  }
-
-  if (!content?.trim() && attachments.length === 0) {
-    return { error: "Message must include content or an attachment.", status: 400 }
-  }
-
-  return { error: null as string | null, status: 200 }
-}
+import { validateChannelTypeMessagePolicy, type SupportedMessageChannelType } from "@/lib/messages/channel-type-policy"
 async function getChannelForRead(supabase: ServerSupabaseClient, channelId: string, userId: string) {
   const { data: channel, error: channelError } = await supabase
     .from("channels")
